CLEANUP |#26| Refactor the test cases by creating an abstract testCase class to handle common operations
